fix(api): combining contained
refs. metriport/metriport-internal#1350
